requirements.txt
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.3.0
opencv-python==4.9.0.80
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.4
packaging>=24.0
tensorflow>=2.10.0
transformers>=4.20.0
torch>=1.12.0
